# cmd/api/docs/swagger.yaml
definitions:
...
  utils.AdminSignUp:
    properties:
      Phonenum:
        type: string
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
    required:
    - Phonenum
    - email
    - firstname
    - lastname
    - password
    type: object
  utils.LoginBody:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  utils.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: MAANUSHI_EARTH_E-COMMERCE REST API built using Go, PSQL, REST API following
    Clean Architecture.
  title: MAANUSHI_EARTH_E-COMMERCE REST API
  version: "2.0"
paths:
...
  /admin/home:
    get:
      consumes:
      - application/json
      description: DISPLAY ADMIN PROFILE
      operationId: ADMIN-PROFILE
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR ADMIN PROFILE
      tags:
      - ADMIN
  /admin/login:
    post:
      consumes:
      - application/json
      description: VERIFY THE EMAIL,PASSWORD, HASH THE PASSWORD AND GENERATE A JWT
        TOKEN AND SET IT TO A COOKIE
      operationId: ADMIN-LOGIN
      parameters:
      - description: Enter the email and password
        in: body
        name: login_details
        required: true
        schema:
          $ref: '#/definitions/utils.LoginBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR ADMIN LOGIN
      tags:
      - ADMIN
  /admin/logout:
    post:
      consumes:
      - application/json
      description: ADMIN LOGOUT
      operationId: ADMIN-LOGOUT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR ADMIN LOGOUT
      tags:
      - ADMIN
...
  /admin/signup:
    post:
      consumes:
      - application/json
      description: CREATE A NEW ADMIN WITH REQUIRED DETAILS
      operationId: SIGNUP-ADMIN
      parameters:
      - description: New Admin Details
        in: body
        name: admin_details
        required: true
        schema:
          $ref: '#/definitions/utils.AdminSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR NEW USER SIGN UP
      tags:
      - ADMIN
  /admin/user:
    get:
      consumes:
      - application/json
      description: LISTING ALL EXISTING USERS
      operationId: ADMIN-LIST-USERS
      parameters:
      - description: Enter the page number to display
        in: query
        name: page
        type: integer
      - description: Number of items to retrieve per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR LISTING USERS
      tags:
      - ADMIN
  /admin/user/{user_id}/make:
    patch:
      consumes:
      - application/json
      description: GRANTING ACCESS FOR INDIVIDUAL USERS.
      operationId: ADMIN-ACCESS
      parameters:
      - description: Enter the specific user id
        in: path
        name: user_id
        required: true
        type: string
      - description: Enter true/false
        in: query
        name: access
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: API FOR BLOCKING/UNBLOCKING USERS
      tags:
      - ADMIN
...
swagger: "2.0"